feat: move node version check to CLI init
